+++ src/components/ChatInterface.tsx
@@ -10,13 +10,11 @@ import { useState, useEffect, useRef } from 'react';
 import { useTheme } from '@/contexts/ThemeContext';
 import { usePersonality } from '@/contexts/PersonalityContext';
 import { PERSONALITIES, PersonalityKey } from '@/constants/personalities';
-import ImageGenerationApproval from './ImageGenerationApproval';
-import StructuredOutputComponent from './StructuredOutput';
-import { parseToolResponse } from '@/lib/structured-parser';
 import SuggestionChips from './SuggestionChips';
 import { useAssistantStatus } from '@/hooks/useAssistantStatus';
 import { useImageGenerationStatus } from '@/hooks/useImageGenerationStatus';
 import { useAutoScroll } from '@/hooks/useAutoScroll';
+import MessagesList from './messages/MessagesList';
 
 interface ChatInterfaceProps {
     chatId: string;
@@ -124,336 +122,13 @@ export default function ChatInterface({ chatId, initialMessages, onNewChat }: Ch
                     backgroundColor: currentTheme.colors.background,
                 }}
             >
-                {error && (
-                    <div className="text-center py-4 px-4 mx-4 mb-4 rounded-lg border" style={{
-                        backgroundColor: '#fef2f2',
-                        borderColor: '#fecaca',
-                        color: '#dc2626'
-                    }}>
-                        <p className="font-medium">Error: {error.message}</p>
-                    </div>
-                )}
-
-                {messages.length === 0 && !error && (
-                    <div className="text-center py-12" style={{ color: currentTheme.colors.text }}>
-                        <p>Ask me for a dad joke!</p>
-                    </div>
-                )}
-
-                {messages?.map((message: any) => (
-                    <div key={message.id} className={`flex ${message.role === 'user' ? 'justify-end' : 'justify-start'}`}>
-                        <div
-                            className="max-w-lg px-4 py-3 rounded-2xl shadow-sm border"
-                            style={message.role === 'user' ? {
-                                backgroundColor: currentTheme.colors.componentColor,
-                                color: currentTheme.colors.text,
-                                borderColor: currentTheme.colors.border
-                            } : {
-                                backgroundColor: currentTheme.colors.surface,
-                                color: currentTheme.colors.text,
-                                borderColor: currentTheme.colors.border
-                            }}
-                        >
-                            <div className="flex items-center gap-2 mb-2">
-                                <div
-                                    className="w-2 h-2 rounded-full"
-                                    style={{
-                                        backgroundColor: message.role === 'user'
-                                            ? 'rgba(255,255,255,0.7)'
-                                            : '#10b981'
-                                    }}
-                                />
-                                <div className="text-xs font-medium opacity-80">
-                                    {message.role === 'user' ? 'You' : PERSONALITIES[personality].label}
-                                </div>
-                            </div>
-
-                            {/* Render message parts */}
-                            <div className="text-sm leading-relaxed whitespace-pre-wrap">
-                                {(() => {
-                                    // Check if this message has a structured tool result - if so, suppress text content AFTER the tool
-                                    const movieSearchPartIndex = message.parts.findIndex((part: any) =>
-                                        part.type === 'tool-movie_search' && part.state === 'output-available'
-                                    );
-                                    const redditPartIndex = message.parts.findIndex((part: any) =>
-                                        part.type === 'tool-reddit' && part.state === 'output-available'
-                                    );
-                                    const youtubePartIndex = message.parts.findIndex((part: any) =>
-                                        part.type === 'tool-youtubeTranscriber' && part.state === 'output-available'
-                                    );
-                                    const websitePartIndex = message.parts.findIndex((part: any) =>
-                                        part.type === 'tool-websiteScraper' && part.state === 'output-available'
-                                    );
-                                    const imageGenerationPartIndex = message.parts.findIndex((part: any) =>
-                                        part.type === 'tool-generate_image' && part.state === 'output-available'
-                                    );
-
-                                    const renderedParts = message.parts.map((part: any, index: number) => {
-                                        switch (part.type) {
-                                            case 'text':
-                                                // Suppress text content if it comes AFTER any structured tool result
-                                                if ((movieSearchPartIndex !== -1 && index > movieSearchPartIndex) ||
-                                                    (redditPartIndex !== -1 && index > redditPartIndex) ||
-                                                    (youtubePartIndex !== -1 && index > youtubePartIndex) ||
-                                                    (websitePartIndex !== -1 && index > websitePartIndex) ||
-                                                    (imageGenerationPartIndex !== -1 && index > imageGenerationPartIndex)) {
-                                                    return null;
-                                                }
-                                                return part.text ? <span key={index}>{part.text}</span> : null;
-
-                                            case 'tool-dad_joke':
-                                                switch (part.state) {
-                                                    case 'input-streaming':
-                                                    case 'input-available':
-                                                    case 'output-available':
-                                                        // Let fallback loading text handle all these states
-                                                        return null;
-                                                    case 'output-error':
-                                                        return (
-                                                            <div key={index} className="text-red-500">
-                                                                Error: {part.errorText}
-                                                            </div>
-                                                        );
-                                                }
-                                                break;
-
-                                            case 'tool-movie_search':
-                                                switch (part.state) {
-                                                    case 'input-streaming':
-                                                    case 'input-available':
-                                                        return (
-                                                            <div key={index} className="text-xs opacity-75 italic">
-                                                                Searching for movies...
-                                                            </div>
-                                                        );
-                                                    case 'output-available':
-                                                        // Parse the structured response and render it
-                                                        const structuredOutput = parseToolResponse('movie_search', part.output as string);
-                                                        if (structuredOutput) {
-                                                            return (
-                                                                <div key={index}>
-                                                                    <StructuredOutputComponent output={structuredOutput} />
-                                                                </div>
-                                                            );
-                                                        }
-                                                        // Fallback to plain text if parsing fails
-                                                        return (
-                                                            <div key={index} className="text-sm">
-                                                                {part.output}
-                                                            </div>
-                                                        );
-                                                    case 'output-error':
-                                                        return (
-                                                            <div key={index} className="text-red-500">
-                                                                Error: {part.errorText}
-                                                            </div>
-                                                        );
-                                                }
-                                                break;
-
-                                            case 'tool-reddit':
-                                                switch (part.state) {
-                                                    case 'input-streaming':
-                                                    case 'input-available':
-                                                        return (
-                                                            <div key={index} className="text-xs opacity-75 italic">
-                                                                Fetching Reddit posts...
-                                                            </div>
-                                                        );
-                                                    case 'output-available':
-                                                        // Parse the structured response and render it
-                                                        const redditOutput = parseToolResponse('reddit', part.output as string);
-                                                        if (redditOutput) {
-                                                            return (
-                                                                <div key={index}>
-                                                                    <StructuredOutputComponent output={redditOutput} />
-                                                                </div>
-                                                            );
-                                                        }
-                                                        // Fallback to plain text if parsing fails
-                                                        return (
-                                                            <div key={index} className="text-sm">
-                                                                {part.output}
-                                                            </div>
-                                                        );
-                                                    case 'output-error':
-                                                        return (
-                                                            <div key={index} className="text-red-500">
-                                                                Error: {part.errorText}
-                                                            </div>
-                                                        );
-                                                }
-                                                break;
-
-                                            case 'tool-youtubeTranscriber':
-                                                switch (part.state) {
-                                                    case 'input-streaming':
-                                                    case 'input-available':
-                                                        return (
-                                                            <div key={index} className="text-xs opacity-75 italic">
-                                                                Analyzing YouTube video transcript...
-                                                            </div>
-                                                        );
-                                                    case 'output-available':
-                                                        // Parse the structured response and render it
-                                                        const youtubeOutput = parseToolResponse('youtubeTranscriber', part.output as string);
-                                                        if (youtubeOutput) {
-                                                            return (
-                                                                <div key={index}>
-                                                                    <StructuredOutputComponent output={youtubeOutput} />
-                                                                </div>
-                                                            );
-                                                        }
-                                                        // Fallback to plain text if parsing fails
-                                                        return (
-                                                            <div key={index} className="text-sm">
-                                                                {part.output}
-                                                            </div>
-                                                        );
-                                                    case 'output-error':
-                                                        return (
-                                                            <div key={index} className="text-red-500">
-                                                                Error: {part.errorText}
-                                                            </div>
-                                                        );
-                                                }
-                                                break;
-
-                                            case 'tool-websiteScraper':
-                                                switch (part.state) {
-                                                    case 'input-streaming':
-                                                    case 'input-available':
-                                                        return (
-                                                            <div key={index} className="text-xs opacity-75 italic">
-                                                                Analyzing website content...
-                                                            </div>
-                                                        );
-                                                    case 'output-available':
-                                                        // Parse the structured response and render it
-                                                        const websiteOutput = parseToolResponse('websiteScraper', part.output as string);
-                                                        if (websiteOutput) {
-                                                            return (
-                                                                <div key={index}>
-                                                                    <StructuredOutputComponent output={websiteOutput} />
-                                                                </div>
-                                                            );
-                                                        }
-                                                        // Fallback to plain text if parsing fails
-                                                        return (
-                                                            <div key={index} className="text-sm">
-                                                                {part.output}
-                                                            </div>
-                                                        );
-                                                    case 'output-error':
-                                                        return (
-                                                            <div key={index} className="text-red-500">
-                                                                Error: {part.errorText}
-                                                            </div>
-                                                        );
-                                                }
-                                                break;
-
-                                            case 'tool-generate_image':
-                                                const callId = part.toolCallId;
-
-                                                switch (part.state) {
-                                                    case 'input-streaming':
-                                                        return (
-                                                            <div key={index} className="text-xs opacity-75 italic">
-                                                                Preparing image request...
-                                                            </div>
-                                                        );
-                                                    case 'input-available':
-                                                        // Show approval UI
-                                                        return (
-                                                            <ImageGenerationApproval
-                                                                key={index}
-                                                                prompt={part.input.prompt}
-                                                                onApprove={() => {
-                                                                    console.log('✅ User approved image generation');
-                                                                    addToolResult({
-                                                                        tool: 'generate_image',
-                                                                        toolCallId: callId,
-                                                                        output: 'APPROVED'
-                                                                    });
-                                                                }}
-                                                                onDeny={() => {
-                                                                    console.log('❌ User denied image generation');
-                                                                    addToolResult({
-                                                                        tool: 'generate_image',
-                                                                        toolCallId: callId,
-                                                                        output: 'DENIED'
-                                                                    });
-                                                                }}
-                                                            />
-                                                        );
-                                                    case 'output-available':
-                                                        // Show the generated image or approval result
-                                                        return (
-                                                            <div key={index}>
-                                                                {part.output === 'APPROVED' || part.output === 'DENIED' ? (
-                                                                    <div className="text-xs opacity-75 italic">
-                                                                        {part.output === 'APPROVED' ? '✅ Generating image...' : '❌ Image generation cancelled'}
-                                                                    </div>
-                                                                ) : (
-                                                                    // Parse the structured response and render it
-                                                                    (() => {
-                                                                        try {
-                                                                            console.log('🔍 Image generation output:', part.output);
-
-                                                                            // Check if it's already a structured object
-                                                                            if (typeof part.output === 'object' && part.output.type === 'image_generation') {
-                                                                                return <StructuredOutputComponent output={part.output} />;
-                                                                            }
-
-                                                                            // Otherwise try to parse as string
-                                                                            const imageOutput = parseToolResponse('generate_image', part.output as string);
-                                                                            if (imageOutput) {
-                                                                                return <StructuredOutputComponent output={imageOutput} />;
-                                                                            }
-                                                                        } catch (error) {
-                                                                            console.error('Error parsing image generation output:', error);
-                                                                        }
-                                                                        // Fallback to plain image if parsing fails
-                                                                        return <img src={part.output as string} alt="Generated" className="rounded-lg max-w-full" />;
-                                                                    })()
-                                                                )}
-                                                            </div>
-                                                        );
-                                                    case 'output-error':
-                                                        return (
-                                                            <div key={index} className="text-red-500">
-                                                                Error: {part.errorText}
-                                                            </div>
-                                                        );
-                                                }
-                                                break;
-
-                                            default:
-                                                return null;
-                                        }
-                                    }).filter(Boolean);
-
-                                    // If nothing to show, display loading text with typing dots
-                                    if (renderedParts.length === 0 && message.role === 'assistant') {
-                                        return (
-                                            <div className="text-xs opacity-75 italic flex items-center gap-1">
-                                                <span>{PERSONALITIES[personality].loadingText}</span>
-                                                <span className="flex gap-0.5">
-                                                    <span className="animate-pulse" style={{ animationDelay: '0s' }}>.</span>
-                                                    <span className="animate-pulse" style={{ animationDelay: '0.2s' }}>.</span>
-                                                    <span className="animate-pulse" style={{ animationDelay: '0.4s' }}>.</span>
-                                                </span>
-                                            </div>
-                                        );
-                                    }
-
-                                    return renderedParts;
-                                })()}
-                            </div>
-                        </div>
-                    </div>
-                ))}
+                <MessagesList
+                    messages={messages}
+                    personality={personality}
+                    currentTheme={currentTheme}
+                    addToolResult={addToolResult}
+                    error={error || null}
+                />
 
                 {/* Scroll anchor */}
                 <div ref={messagesEndRef} style={{ marginBottom: '20px' }} />
